backend/requirements.txt
# FastAPI & Web Server
fastapi==0.110.0
uvicorn[standard]==0.28.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.28
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.2
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation & Serialization
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.1

# File Upload
boto3==1.34.58
python-magic==0.4.27
pillow==10.2.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0
faker==24.2.0

# Development
black==24.2.0
isort==5.13.2
mypy==1.9.0
ruff==0.3.2

